create a new action
